--- src/main/java/com/sponus/sponusbe/domain/propose/controller/ProposeCreateResponse.java
@@ -1,6 +1,0 @@
-package com.sponus.sponusbe.domain.propose.controller;
-
-public record ProposeCreateResponse(
-	Long proposeId
-) {
-}
