--- src/main/java/com/sponus/sponusbe/domain/propose/controller/ProposeType.java
@@ -1,6 +1,0 @@
-package com.sponus.sponusbe.domain.propose.controller;
-
-public enum ProposeType {
-	SEND, RECEIVED;
-
-}
