+++ src/main/java/com/sponus/sponusbe/domain/propose/controller/dto/request/ProposeType.java
@@ -1,0 +1,6 @@
+package com.sponus.sponusbe.domain.propose.controller.dto.request;
+
+public enum ProposeType {
+	SEND, RECEIVED;
+
+}
