+++ src/main/java/com/sponus/sponusbe/domain/propose/controller/dto/response/ProposeCreateResponse.java
@@ -1,0 +1,6 @@
+package com.sponus.sponusbe.domain.propose.controller.dto.response;
+
+public record ProposeCreateResponse(
+	Long proposeId
+) {
+}
